Test: Allow tests to run in parallel.
It's worth noting that this doesn't actually work. The tests fail. This
has highlighted a number of weird quirks in our controller when one
test cleans up at just the wrong time. Containers go missing, 404s go
flying, calamity ensues.

I think these problems will be addressed by making harvesting a
background job, so we can leave it here and maybe switch to this mode as
the default when it's stable.

Here's an example of the errors we might get running tests in parallel:

    [2018-04-18 18:53:14,358] ERROR in app: Exception on /executions/bd30ac06-4339-11e8-8805-6bbc9a67228c/logs [GET]
    Traceback (most recent call last):
      File "/usr/local/lib/python3.6/site-packages/docker/api/client.py", line 225, in _raise_for_status
        response.raise_for_status()
      File "/usr/local/lib/python3.6/site-packages/requests/models.py", line 935, in raise_for_status
        raise HTTPError(http_error_msg, response=self)
    requests.exceptions.HTTPError: 404 Client Error: Not Found for url: http+docker://localhost/v1.35/containers/f3819193e947894fa0e54a202ba9ac6ec45c6ee0cffc55a3fab25e4b15052c0b/json