Do not execute startup.yml in workers
It basically sets up a cache. In the worker we are discarding
the cache so it's useless. We want to implement some sort of
cache eventually but the right way to do it is to create snapshots
with /var/lib/docker dirs with the images, and attach them to workers
(having different stuff in the cache every time goes against
reproducibility, and user code in the images costs usually nothing to
transfer)