Make the
Previous to this there was a hacky hack to create a symlink
to the controller code if it didn't exist. This worked as to run
the `plz` script to execute the code directly, but was breaking
`make check` for the cli, as the pipenv for the cli lacked the
modules required by the controller. If we have a link to the
controller api instead, everything works fine.

Underlying all this is the fact that if there are several
paths to a module in the pythonpath, the python will pick only
one of them (as opposed to java). So we cannot put the api
in a directory outside the controller and the api code under
plz.controller.api (as we'd do in java) as only either the controller
or the api would be picked