plz-output: improve performance and reliability
- Removed lock timeout. For outputs taking more than one minute
to download, it was timing out and then failing when releasing the
lock. Downloading in the meantime could cause getting partial
output
- Improved how long responses are read from the CLI. It was reading small
amounts at a time, implying a lot of writes, a huge waste of time.
This might be due to a change in the underlying libraries, as we
were having big outputs before and this wasn't a problem